+++ javascript/node/selenium-webdriver/io/index.js
@@ -14,13 +14,119 @@
 // limitations under the License.
 
 var fs = require('fs'),
-    path = require('path');
+    ncp = require('ncp').ncp,
+    path = require('path'),
+    tmp = require('tmp');
+
+var promise = require('..').promise;
 
 
 var PATH_SEPARATOR = process.platform === 'win32' ? ';' : ':';
 
 
+/**
+ * Creates the specified directory and any necessary parent directories. No
+ * action is taken if the directory already exists.
+ * @param {string} dir The directory to create.
+ * @param {function(Error)} callback Callback function; accepts a single error
+ *     or {@code null}.
+ */
+function createDirectories(dir, callback) {
+  fs.mkdir(dir, function(err) {
+    if (!err || err.code === 'EEXIST') {
+      callback();
+    } else {
+      createDirectories(path.dirname(dir), function(err) {
+        err && callback(err) || fs.mkdir(dir, callback);
+      });
+    }
+  });
+};
+
 // PUBLIC API
+
+
+
+/**
+ * Copies one file to another.
+ * @param {string} src The source file.
+ * @param {string} dst The destination file.
+ * @return {!promise.Promise.<string>} A promise for the copied file's path.
+ */
+exports.copy = function(src, dst) {
+  var copied = promise.defer();
+
+  var rs = fs.createReadStream(src);
+  rs.on('error', copied.reject);
+  rs.on('end', function() {
+    copied.fulfill(dst);
+  });
+
+  var ws = fs.createWriteStream(dst);
+  ws.on('error', copied.reject);
+
+  rs.pipe(ws);
+
+  return copied.promise;
+};
+
+
+/**
+ * Recursively copies the contents of one directory to another.
+ * @param {string} src The source directory to copy.
+ * @param {string} dst The directory to copy into.
+ * @param {(RegEx|function(string): boolean)=} opt_exclude An exclusion filter
+ *     as either a regex or predicate function. All files matching this filter
+ *     will not be copied.
+ * @return {!promise.Promise.<string>} A promise for the destination
+ *     directory's path once all files have been copied.
+ */
+exports.copyDir = function(src, dst, opt_exclude) {
+  var predicate = opt_exclude;
+  if (opt_exclude && typeof opt_exclude !== 'function') {
+    predicate = function(p) {
+      return !opt_exclude.test(p);
+    };
+  }
+
+  var copied = promise.defer();
+  ncp(src, dst, {filter: predicate}, function(err) {
+    err && copied.reject(err) || copied.fulfill(dst);
+  });
+  return copied.promise;
+};
+
+
+/**
+ * Tests if a file path exists.
+ * @param {string} path The path to test.
+ * @return {!promise.Promise.<boolean>} A promise for whether the file exists.
+ */
+exports.exists = function(path) {
+  var result = promise.defer();
+  fs.exists(path, result.fulfill);
+  return result.promise;
+};
+
+
+/**
+ * @return {!promise.Promise.<string>} A promise for the path to a temporary
+ *     directory.
+ * @see https://www.npmjs.org/package/tmp
+ */
+exports.tmpDir = function() {
+  return promise.checkedNodeCall(tmp.dir);
+};
+
+
+/**
+ * @return {!promise.Promise.<string>} A promise for the path to a temporary
+ *     file.
+ * @see https://www.npmjs.org/package/tmp
+ */
+exports.tmpFile = function() {
+  return promise.checkedNodeCall(tmp.file);
+};
 
 
 /**